Fix build: move Basic auth into proxy.ts, remove middleware.ts
Nextra prohibits having both middleware.ts and proxy.ts. Merged the
blog preview Basic auth logic into proxy.ts which is Nextra's
middleware extension point.
