+++ src/proxy.ts
@@ -3,12 +3,44 @@ import { type NextRequest, NextResponse } from "next/server";
 const locales = ["en", "ja"];
 const defaultLocale = "en";
 
+const PREVIEW_PASSWORD = process.env.BLOG_PREVIEW_PASSWORD ?? "";
+
+function requireBasicAuth(request: NextRequest): NextResponse | null {
+	if (!PREVIEW_PASSWORD) {
+		return new NextResponse("Preview not available: BLOG_PREVIEW_PASSWORD not set", {
+			status: 503,
+		});
+	}
+	const authHeader = request.headers.get("authorization");
+	if (!authHeader?.startsWith("Basic ")) {
+		return new NextResponse("Unauthorized", {
+			status: 401,
+			headers: { "WWW-Authenticate": 'Basic realm="Blog Preview"' },
+		});
+	}
+	const decoded = atob(authHeader.slice(6));
+	const pass = decoded.slice(decoded.indexOf(":") + 1);
+	if (pass !== PREVIEW_PASSWORD) {
+		return new NextResponse("Unauthorized", {
+			status: 401,
+			headers: { "WWW-Authenticate": 'Basic realm="Blog Preview"' },
+		});
+	}
+	return null;
+}
+
 function getLocale(_request: NextRequest) {
 	return defaultLocale;
 }
 
 export function proxy(request: NextRequest) {
 	const { pathname } = request.nextUrl;
+
+	if (/^\/(en|ja)\/blog\/preview\//.test(pathname)) {
+		const authResponse = requireBasicAuth(request);
+		if (authResponse) return authResponse;
+	}
+
 	const pathnameHasLocale = locales.some(
 		(locale) => pathname.startsWith(`/${locale}/`) || pathname === `/${locale}`,
 	);
@@ -48,6 +80,7 @@ export function proxy(request: NextRequest) {
 	}
 	return NextResponse.redirect(request.nextUrl);
 }
+
 export const config = {
 	matcher: [
 		// Skip all internal paths (_next)
